Refactor: update bind cards component structure; replace PublicBindsCards with BindCards; enhance data handling for user binds; adjust import paths for consistency
