+++ next/src/components/layout/binds-cards/public-binds-card.tsx
