+++ next/src/types/auth/user.ts
@@ -1,5 +1,6 @@
 type IBinds = {
   name: string;
+  code: string;
   description: string;
   createdAt: string;
   updatedAt: string;
